refactor: rename schemas.py to currency.py and move to a new folder named schemas/
